refactor(templates): move custom templates into templates
